fix: resolve choreoId JSON parsing & enforce absolute black text on dancer tags

+++ src/jsMain/kotlin/com/licencjat/licencjat_frontend/ApiService.kt
@@ -136,7 +136,7 @@ object ApiService {
         }
     }
 
-    fun createTask(title: String, desc: String, deadline: String, choreoId: Long, dancerIds: List<String>, file: dynamic): kotlin.js.Promise<dynamic> {
+    fun createTask(title: String, desc: String, deadline: String, choreoId: Int, dancerIds: List<String>, file: dynamic): kotlin.js.Promise<dynamic> {
         // Krok 1: upload pliku bezpośrednio do Cloudinary z przeglądarki
         return uploadToCloudinary(file).then<dynamic> { cloudinaryRes: dynamic ->
             val videoUrl = cloudinaryRes.secure_url?.toString()
